authentication and authorization using api

+++ controllers/api/v1/posts_api.js
@@ -22,18 +22,19 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "Post and associated coments deleted successfully",
-    });
-    // } else {
-    //   req.flash("error", "You cannot delete this post!");
-    //   return res.redirect("back");
-    // }
+      return res.json(200, {
+        message: "Post and associated coments deleted successfully",
+      });
+    } else {
+      return req.json(401, {
+        message: "You cannot delete this post!",
+      });
+    }
   } catch (err) {
     console.log(err);
     return res.json(500, {
